// apps/web-svelte/src/lib/api/generated/model/entitySharedSourceEpisode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface EntitySharedSourceEpisode {
  id: string;
  title: string;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  seasonNumber: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  episodeNumber: number | string | null;
}